+++ azure_devops_rust_api/examples/pipeline_preview.rs
@@ -1,11 +1,11 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-// pipelines.rs
-// Pipelines example.
+// pipeline_preview.rs
+// Pipeline preview example.
 use anyhow::Result;
 use azure_devops_rust_api::pipelines;
-use azure_devops_rust_api::pipelines::models::Pipeline;
+use azure_devops_rust_api::pipelines::models::{Pipeline, RunPipelineParameters};
 use azure_devops_rust_api::Credential;
 use std::env;
 use std::sync::Arc;
@@ -29,7 +29,7 @@ async fn main() -> Result<()> {
         env::var("ADO_SERVICE_ENDPOINT").expect("Must define ADO_SERVICE_ENDPOINT");
     let organization = env::var("ADO_ORGANIZATION").expect("Must define ADO_ORGANIZATION");
     let project = env::var("ADO_PROJECT").expect("Must define ADO_PROJECT");
-    let pipeline_name = env::args().nth(1).expect("Usage: pipelines <name>");
+    let pipeline_name = env::args().nth(1).expect("Usage: pipeline_preview <name>");
 
     // Create a `pipelines` client
     let client = pipelines::operations::Client::new(service_endpoint, credential, vec![]);
@@ -51,42 +51,30 @@ async fn main() -> Result<()> {
         .cloned()
         .collect();
 
-    for pipeline in matched_pipelines.iter() {
-        println!("{:4} {}", pipeline.id, pipeline.name);
-    }
-
     if let Some(pipeline) = matched_pipelines.iter().next() {
-        println!("\nExample pipeline struct from list:");
-        println!("{:#?}", pipeline);
+        // Demonstrate how to query a preview of pipeline YAML...
+        // Define the pipeline params
+        let run_pipeline_params = RunPipelineParameters {
+            preview_run: Some(true),
+            resources: None,
+            stages_to_skip: vec![],
+            template_parameters: None,
+            variables: None,
+            yaml_override: None,
+        };
 
-        // The pipeline struct returned from list is different from that returned by get.
-        // Query and display the struct returned by get for comparison.
-        let pipeline = client
-            .pipelines()
-            .get(&organization, &project, pipeline.id)
+        // Create a preview client
+        let preview_client = client.preview();
+
+        // Request a preview of the specified pipeline
+        let preview = preview_client
+            .preview(&organization, run_pipeline_params, &project, pipeline.id)
             .into_future()
             .await?;
-        println!("\nExample pipeline struct from get:");
-        println!("{:#?}", pipeline);
 
-        // Use the client to list all runs of the selected pipeline
-        let runs = client
-            .runs()
-            .list(&organization, &project, pipeline.id)
-            .into_future()
-            .await?
-            .value;
-
-        println!("\nPipeline runs: {}", runs.len());
-        // Display [result, state] for each pipeline run
-        for run in runs.iter() {
-            println!(
-                "{:8} {:16} {:14} {:14}",
-                run.run_reference.id,
-                run.run_reference.name,
-                format!("{:?}", run.result),
-                format!("{:?}", run.state)
-            );
+        // Display the full pipeline YAML
+        if let Some(final_yaml) = preview.final_yaml {
+            println!("Pipeline preview:\n{}", final_yaml);
         }
     }